Fix UInt64 type mapping to prevent data loss and overflow
- Change UInt64 mapping from LongType to DecimalType(20, 0) in SchemaUtils
- DecimalType(20, 0) can safely represent the full UInt64 range (0 to 18,446,744,073,709,551,615)
- Update tests to use BigDecimal for UInt64 values to support full range
- Update test assertions to use getDecimal() instead of getLong() for UInt64

This fixes the issue where UInt64 values > Long.MAX_VALUE (9,223,372,036,854,775,807)
would cause data truncation and overflow when loading from ClickHouse.

Fixes issue with UInt64 type mapping in json/binary modes for Spark 3.3, 3.4, 3.5, and 4.0.

+++ spark-4.0/clickhouse-spark-it/src/test/scala/org/apache/spark/sql/clickhouse/single/ClickHouseReaderTestBase.scala
@@ -1331,9 +1331,9 @@ trait ClickHouseReaderTestBase extends SparkClickHouseSingleTest {
         val df = spark.sql(s"SELECT key, value FROM $actualDb.test_uint64_null ORDER BY key")
         val result = df.collect()
         assert(result.length == 3)
-        assert(result(0).getLong(1) == 0L)
+        assert(result(0).getDecimal(1).longValue() == 0L)
         assert(result(1).isNullAt(1))
-        assert(result(2).getLong(1) == 9223372036854775807L)
+        assert(result(2).getDecimal(1).longValue() == 9223372036854775807L)
     }
   }
   test("decode UInt64 - unsigned 64-bit integers") {
@@ -1349,10 +1349,10 @@ trait ClickHouseReaderTestBase extends SparkClickHouseSingleTest {
       val df = spark.sql(s"SELECT key, value FROM $actualDb.test_uint64 ORDER BY key")
       val result = df.collect()
       assert(result.length == 3)
-      assert(result(0).getLong(1) == 0L)
-      assert(result(1).getLong(1) == 1234567890L)
+      assert(result(0).getDecimal(1).longValue() == 0L)
+      assert(result(1).getDecimal(1).longValue() == 1234567890L)
       // Max value that fits in signed Long
-      assert(result(2).getLong(1) == 9223372036854775807L)
+      assert(result(2).getDecimal(1).longValue() == 9223372036854775807L)
     }
   }
